+++ src/components/edit/__tests__/edit-test-cases.js
@@ -44,8 +44,7 @@ testCases.noCss = {
       title: 'Display a map'
     },
     head: `<meta name='viewport' content='initial-scale=1,maximum-scale=1,user-scalable=no' />`,
-    js:
-      "var map = new mapboxgl.Map({\n    container: 'map', // container id\n    style: 'mapbox://styles/mapbox/streets-v11', // stylesheet location\n    center: [-74.50, 40], // starting position [lng, lat]\n    zoom: 9 // starting zoom\n});",
+    js: "var map = new mapboxgl.Map({\n    container: 'map', // container id\n    style: 'mapbox://styles/mapbox/streets-v11', // stylesheet location\n    center: [-74.50, 40], // starting position [lng, lat]\n    zoom: 9 // starting zoom\n});",
     resources: {
       js: ['https://api.tiles.mapbox.com/mapbox-gl-js/v1.5.0/mapbox-gl.js'],
       css: ['https://api.tiles.mapbox.com/mapbox-gl-js/v1.5.0/mapbox-gl.css']